Fix sign-compare warning

+++ ext/src/swow_hook.c
@@ -31,7 +31,7 @@ static cat_bool_t swow_function_is_hookable(const char *name, size_t name_length
 	while (1) {
 		if (*e == ' ' || *e == ',' || *e == '\0') {
 			if (s) {
-                if (e - s == name_length && cat_strncasecmp(s, name, name_length) == 0) {
+                if ((size_t) (e - s) == name_length && cat_strncasecmp(s, name, name_length) == 0) {
                     return cat_false;
                 }
 				s = NULL;
